Add analytics permission switch to settings
