+++ app/src/main/java/no/nordicsemi/android/nrftoolbox/view/FeatureButton.kt
@@ -20,8 +20,8 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import no.nordicsemi.android.nrftoolbox.R
 import no.nordicsemi.android.material.you.ScreenSection
+import no.nordicsemi.android.nrftoolbox.R
 
 @Composable
 fun FeatureButton(
