+++ app/src/main/java/no/nordicsemi/android/nrftoolbox/view/TitleAppBar.kt
@@ -1,0 +1,27 @@
+package no.nordicsemi.android.nrftoolbox.view
+
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.SmallTopAppBar
+import androidx.compose.material3.Text
+import androidx.compose.material3.TopAppBarDefaults
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.res.colorResource
+import no.nordicsemi.analytics.view.AnalyticsPermissionButton
+import no.nordicsemi.android.theme.R
+
+@Composable
+fun TitleAppBar(text: String) {
+    SmallTopAppBar(
+        title = { Text(text, maxLines = 2) },
+        colors = TopAppBarDefaults.smallTopAppBarColors(
+            scrolledContainerColor = MaterialTheme.colorScheme.primary,
+            containerColor = colorResource(id = R.color.appBarColor),
+            titleContentColor = MaterialTheme.colorScheme.onPrimary,
+            actionIconContentColor = MaterialTheme.colorScheme.onPrimary,
+            navigationIconContentColor = MaterialTheme.colorScheme.onPrimary,
+        ),
+        actions = {
+            AnalyticsPermissionButton()
+        }
+    )
+}
