+++ app/src/main/java/no/nordicsemi/android/nrftoolbox/viewmodel/HomeViewModel.kt
@@ -1,6 +1,5 @@
 package no.nordicsemi.android.nrftoolbox.viewmodel
 
-import android.util.Log
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
