+++ lib_analytics/build.gradle
@@ -8,4 +8,6 @@ dependencies {
     implementation platform('com.google.firebase:firebase-bom:29.2.1')
     implementation 'com.google.firebase:firebase-analytics'
     implementation 'com.google.firebase:firebase-crashlytics'
+
+    implementation libs.nordic.analytics
 }
