+++ lib_analytics/src/main/java/no/nordicsemi/android/analytics/AppAnalytics.kt
@@ -1,22 +1,17 @@
 package no.nordicsemi.android.analytics
 
 import android.annotation.SuppressLint
-import android.content.Context
-import com.google.firebase.analytics.FirebaseAnalytics
-import dagger.hilt.android.qualifiers.ApplicationContext
+import no.nordicsemi.analytics.NordicAnalytics
 import javax.inject.Inject
 import javax.inject.Singleton
 
 @SuppressLint("MissingPermission")
 @Singleton
 class AppAnalytics @Inject constructor(
-    @ApplicationContext
-    private val context: Context
+    private val nordicAnalytics: NordicAnalytics
 ) {
 
-    private val firebase by lazy { FirebaseAnalytics.getInstance(context) }
-
     fun logEvent(event: FirebaseEvent) {
-        firebase.logEvent(event.eventName, event.params)
+        nordicAnalytics.logEvent(event.eventName, event.params)
     }
 }
