+++ lib_service/src/main/java/no/nordicsemi/android/service/BleManagerStatus.kt
@@ -27,6 +27,7 @@ sealed class DeviceHolder(val device: BluetoothDevice) {
 
 class IdleResult<T> : BleManagerResult<T>
 class ConnectingResult<T>(device: BluetoothDevice) : DeviceHolder(device), BleManagerResult<T>
+class ConnectedResult<T>(device: BluetoothDevice) : DeviceHolder(device), BleManagerResult<T>
 class SuccessResult<T>(device: BluetoothDevice, val data: T) : DeviceHolder(device), BleManagerResult<T>
 
 class LinkLossResult<T>(device: BluetoothDevice, val data: T?) : DeviceHolder(device), BleManagerResult<T>
