+++ profile_bps/src/main/java/no/nordicsemi/android/bps/view/BPSViewState.kt
@@ -2,7 +2,6 @@ package no.nordicsemi.android.bps.view
 
 import no.nordicsemi.android.bps.data.BPSData
 import no.nordicsemi.android.service.BleManagerResult
-import no.nordicsemi.ui.scanner.DiscoveredBluetoothDevice
 
 internal sealed class BPSViewState
