+++ profile_cgms/src/main/java/no/nordicsemi/android/cgms/repository/CGMService.kt
@@ -1,6 +1,5 @@
 package no.nordicsemi.android.cgms.repository
 
-import android.bluetooth.BluetoothDevice
 import android.content.Intent
 import androidx.lifecycle.lifecycleScope
 import dagger.hilt.android.AndroidEntryPoint
