+++ profile_cgms/src/main/java/no/nordicsemi/android/cgms/view/CGMContentView.kt
@@ -5,6 +5,7 @@ import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Search
 import androidx.compose.material.icons.filled.Settings
 import androidx.compose.material3.Button
+import androidx.compose.material3.CircularProgressIndicator
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
@@ -17,10 +18,9 @@ import no.nordicsemi.android.cgms.data.CGMData
 import no.nordicsemi.android.cgms.data.CGMRecord
 import no.nordicsemi.android.cgms.data.CGMServiceCommand
 import no.nordicsemi.android.cgms.data.RequestStatus
-import no.nordicsemi.android.theme.view.BatteryLevelView
 import no.nordicsemi.android.material.you.ScreenSection
+import no.nordicsemi.android.theme.view.BatteryLevelView
 import no.nordicsemi.android.theme.view.SectionTitle
-import androidx.compose.material3.CircularProgressIndicator
 
 @Composable
 internal fun CGMContentView(state: CGMData, onEvent: (CGMViewEvent) -> Unit) {
