+++ profile_cgms/src/main/java/no/nordicsemi/android/cgms/viewmodel/CGMViewModel.kt
@@ -9,11 +9,11 @@ import no.nordicsemi.android.analytics.AppAnalytics
 import no.nordicsemi.android.analytics.Profile
 import no.nordicsemi.android.analytics.ProfileConnectedEvent
 import no.nordicsemi.android.cgms.data.CGMS_SERVICE_UUID
-import no.nordicsemi.android.cgms.repository.CGMRepository
 import no.nordicsemi.android.cgms.data.CGMServiceCommand
+import no.nordicsemi.android.cgms.repository.CGMRepository
 import no.nordicsemi.android.cgms.view.*
 import no.nordicsemi.android.navigation.*
-import no.nordicsemi.android.service.SuccessResult
+import no.nordicsemi.android.service.ConnectedResult
 import no.nordicsemi.android.utils.exhaustive
 import no.nordicsemi.android.utils.getDevice
 import no.nordicsemi.ui.scanner.ScannerDestinationId
@@ -39,7 +39,7 @@ internal class CGMViewModel @Inject constructor(
         repository.data.onEach {
             _state.value = WorkingState(it)
 
-            (it as? SuccessResult)?.let {
+            (it as? ConnectedResult)?.let {
                 analytics.logEvent(ProfileConnectedEvent(Profile.CGMS))
             }
         }.launchIn(viewModelScope)
