+++ profile_csc/src/main/java/no/nordicsemi/android/csc/view/SelectWheelSizeDialog.kt
@@ -5,15 +5,11 @@ import androidx.compose.ui.res.stringArrayResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import no.nordicsemi.android.csc.R
-import no.nordicsemi.android.csc.data.WheelSize
 import no.nordicsemi.android.material.you.NordicTheme
-import no.nordicsemi.android.theme.view.dialog.FlowCanceled
-import no.nordicsemi.android.theme.view.dialog.ItemSelectedResult
 import no.nordicsemi.android.theme.view.dialog.StringListDialog
 import no.nordicsemi.android.theme.view.dialog.StringListDialogConfig
 import no.nordicsemi.android.theme.view.dialog.StringListDialogResult
 import no.nordicsemi.android.theme.view.dialog.toAnnotatedString
-import no.nordicsemi.android.utils.exhaustive
 
 @Composable
 internal fun SelectWheelSizeDialog(onEvent: (StringListDialogResult) -> Unit) {
