+++ profile_hrs/src/main/java/no/nordicsemi/android/hrs/service/HRSService.kt
@@ -1,6 +1,5 @@
 package no.nordicsemi.android.hrs.service
 
-import android.bluetooth.BluetoothDevice
 import android.content.Intent
 import androidx.lifecycle.lifecycleScope
 import dagger.hilt.android.AndroidEntryPoint
