+++ profile_hts/src/main/java/no/nordicsemi/android/hts/view/HTSContentView.kt
@@ -12,9 +12,9 @@ import androidx.compose.ui.unit.dp
 import no.nordicsemi.android.hts.R
 import no.nordicsemi.android.hts.data.HTSData
 import no.nordicsemi.android.material.you.RadioButtonGroup
+import no.nordicsemi.android.material.you.ScreenSection
 import no.nordicsemi.android.theme.view.BatteryLevelView
 import no.nordicsemi.android.theme.view.KeyValueField
-import no.nordicsemi.android.material.you.ScreenSection
 import no.nordicsemi.android.theme.view.SectionTitle
 
 @Composable
