+++ profile_prx/src/main/java/no/nordicsemi/android/prx/repository/PRXService.kt
@@ -1,6 +1,5 @@
 package no.nordicsemi.android.prx.repository
 
-import android.bluetooth.BluetoothDevice
 import android.content.Intent
 import androidx.lifecycle.lifecycleScope
 import dagger.hilt.android.AndroidEntryPoint
