+++ profile_uart/src/main/java/no/nordicsemi/android/uart/repository/UARTService.kt
@@ -1,6 +1,5 @@
 package no.nordicsemi.android.uart.repository
 
-import android.bluetooth.BluetoothDevice
 import android.content.Intent
 import androidx.lifecycle.lifecycleScope
 import dagger.hilt.android.AndroidEntryPoint
