+++ profile_uart/src/main/java/no/nordicsemi/android/uart/view/UARTAddMacroDialog.kt
@@ -1,6 +1,8 @@
 package no.nordicsemi.android.uart.view
 
-import androidx.compose.foundation.*
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.lazy.grid.GridCells
 import androidx.compose.foundation.lazy.grid.GridItemSpan
